+++ Options/CityHallOptions.cs
@@ -12,7 +12,7 @@ public class CityHallOptions :
     public int ElectricityConsumption { get; set; }
     public int WaterConsumption { get; set; }
     public int GarbageAccumulation { get; set; }
-    public int TelecomNeed { get; set; }
+    public float TelecomNeed { get; set; }
     public float CityLoanIntrest { get; set; }
     public float CityImportcost { get; set; }
