+++ Options/FirewatchTowerOptions.cs
@@ -12,7 +12,7 @@ public class FirewatchTowerOptions :
     public int ElectricityConsumption { get; set; }
     public int WaterConsumption { get; set; }
     public int GarbageAccumulation { get; set; }
-    public int TelecomNeed { get; set; }
+    public float TelecomNeed { get; set; }
     public float LocalForestFireResponseTime { get; set; }
     public float LocalForestFireResponseTimeRadius { get; set; }
     public float LocalForestFireHazard { get; set; }
