+++ Options/ParkinghallOptions.cs
@@ -13,7 +13,7 @@ public class ParkinghallOptions :
     public int ElectricityConsumption { get; set; }
     public int WaterConsumption { get; set; }
     public int GarbageAccumulation { get; set; }
-    public int TelecomNeed { get; set; }
+    public float TelecomNeed { get; set; }
     public int GroundPollution { get; set; }
     public int AirPollution { get; set; }
     public int NoisePollution { get; set; }
