+++ Options/PoliceCarOptions.cs
@@ -7,7 +7,7 @@ public class PoliceCarOptions :
     public string Name { get; set; } = null!;
 
     public int CriminalCapacity { get; set; }
-    public int CrimeReductionRate { get; set; }
+    public float CrimeReductionRate { get; set; }
 
 
 }
