+++ Options/RadioMastOptions.cs
@@ -14,9 +14,9 @@ public class RadioMastOptions :
     public int ElectricityConsumption { get; set; }
     public int WaterConsumption { get; set; }
     public int GarbageAccumulation { get; set; }
-    public int TelecomNeed { get; set; }
-    public int TelecomRange { get; set; }
-    public int NetworkCapacity { get; set; }
+    public float TelecomNeed { get; set; }
+    public float TelecomRange { get; set; }
+    public float NetworkCapacity { get; set; }
     public int GroundPollution { get; set; }
     public int AirPollution { get; set; }
     public int NoisePollution { get; set; }
