+++ Options/RadioTelescopeOptions.cs
@@ -11,7 +11,7 @@ public class RadioTelescopeOptions :
     public int ElectricityConsumption { get; set; }
     public int WaterConsumption { get; set; }
     public int GarbageAccumulation { get; set; }
-    public int TelecomNeed { get; set; }
+    public float TelecomNeed { get; set; }
     public float CityUniversityGraduation { get; set; }
 
 }
