+++ Options/ServerfarmOptions.cs
@@ -15,12 +15,12 @@ public class ServerfarmOptions :
     public int ElectricityConsumption { get; set; }
     public int WaterConsumption { get; set; }
     public int GarbageAccumulation { get; set; }
-    public int TelecomNeed { get; set; }
+    public float TelecomNeed { get; set; }
     public int GroundPollution { get; set; }
     public int AirPollution { get; set; }
     public int NoisePollution { get; set; }
-    public int TelecomRange { get; set; }
-    public int NetworkCapacity { get; set; }
+    public float TelecomRange { get; set; }
+    public float NetworkCapacity { get; set; }
     public float CityEntertainment { get; set; }
     public float CityTelecomCapacity { get; set; }
     public float CityAttractiveness { get; set; }
